feat: implementado o teste referente ao exercício 4 e ajustado o tipo de retorno, pois o assert do junit não aceita integer por redundancia.

+++ testesUnitarios/src/main/java/easy/Exercicio04_easy.java
@@ -4,8 +4,8 @@ import javax.swing.*;
 
 public class Exercicio04_easy {//Inicio classe
 
-    public Integer calcularDobro(int numeroDigitado) {
-        int dobro = numeroDigitado * 2;
+    public long calcularDobro(long numeroDigitado) {
+        long dobro = numeroDigitado * 2;
         return dobro;
     }
 }
